Move WindowClosedError to the platform contract (extraction plan phase 1)
First leaf cleanup from docs/AGENT-BASE-PLAN.md: generic alert/DM send
paths (router, health, backgroundJobs, usageAlert, usageCostDigest,
backgroundJobCostAlert, departedAdminAlert, agent/tools) instanceof-check
WindowClosedError, so the class belongs on the platform abstraction in
platforms/types.ts — not in the WhatsApp Cloud adapter they otherwise
have no business importing. cloudAdapter re-exports it so existing import
sites (including tests) keep working. No behaviour change.
